# More efficient parsing of valid 36-byte UUIDs