--- a/build.sc
+++ b/build.sc
@@ -38,7 +38,7 @@ trait CommonModule extends ScalaModule with ScalafmtModule {
 }
 
 trait CommonPublishModule extends CommonModule with PublishModule with CrossScalaModule{
-  def publishVersion = "1.0.1"
+  def publishVersion = "1.0.2-SNAPSHOT"
 
   def pomSettings = PomSettings(
     description = artifactName(),