This PR makes parsing of valid 36-byte UUIDs faster in 2x-7x times (depending on the JDK version) without changes in handling of errors and shorter values (like `"0-0-0-0-0"`) or longer  values (yeap, the standard Java parser accepts them).

Bellow are results of benchmarks that parse an array of 128 JSON strings to an array of UUIDs.

## JDK 8
### Before
```
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                             (size)   Mode  Cnt       Score      Error  Units
[info] ArrayOfUUIDsReading.avSystemGenCodec     128  thrpt    5   14134.050 ±  271.126  ops/s
[info] ArrayOfUUIDsReading.borerJson            128  thrpt    5   14066.702 ±  241.422  ops/s
[info] ArrayOfUUIDsReading.circe                128  thrpt    5   12880.728 ±   98.472  ops/s
[info] ArrayOfUUIDsReading.dslJsonScala         128  thrpt    5  129764.835 ± 5411.463  ops/s
[info] ArrayOfUUIDsReading.jacksonScala         128  thrpt    5   60901.845 ±  693.853  ops/s
[info] ArrayOfUUIDsReading.jsoniterScala        128  thrpt    5  232747.513 ± 1583.347  ops/s
[info] ArrayOfUUIDsReading.playJson             128  thrpt    5   12136.594 ±  109.441  ops/s
[info] ArrayOfUUIDsReading.sprayJson            128  thrpt    5    8472.291 ±   30.612  ops/s
[info] ArrayOfUUIDsReading.uPickle              128  thrpt    5   12786.165 ±   83.007  ops/s
[info] ArrayOfUUIDsReading.weePickle            128  thrpt    5   13626.593 ±   59.815  ops/s
```
### After
```
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                             (size)   Mode  Cnt       Score      Error  Units
[info] ArrayOfUUIDsReading.avSystemGenCodec     128  thrpt    5   14120.908 ±  124.210  ops/s
[info] ArrayOfUUIDsReading.borerJson            128  thrpt    5   13506.458 ±  126.836  ops/s
[info] ArrayOfUUIDsReading.circe                128  thrpt    5   12791.404 ±  112.205  ops/s
[info] ArrayOfUUIDsReading.dslJsonScala         128  thrpt    5  127389.261 ± 8247.226  ops/s
[info] ArrayOfUUIDsReading.jacksonScala         128  thrpt    5   60863.041 ±  433.560  ops/s
[info] ArrayOfUUIDsReading.jsoniterScala        128  thrpt    5  232451.966 ±  503.282  ops/s
[info] ArrayOfUUIDsReading.playJson             128  thrpt    5   12231.396 ±  116.315  ops/s
[info] ArrayOfUUIDsReading.sprayJson            128  thrpt    5    8596.765 ±   39.366  ops/s
[info] ArrayOfUUIDsReading.uPickle              128  thrpt    5   12873.315 ±  112.585  ops/s
[info] ArrayOfUUIDsReading.weePickle            128  thrpt    5   78200.411 ±  403.595  ops/s
```

## JDK 15
### Before
```
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                             (size)   Mode  Cnt       Score     Error  Units
[info] ArrayOfUUIDsReading.avSystemGenCodec     128  thrpt    5   46183.155 ± 937.379  ops/s
[info] ArrayOfUUIDsReading.borerJson            128  thrpt    5   44264.778 ± 212.971  ops/s
[info] ArrayOfUUIDsReading.circe                128  thrpt    5   42792.567 ± 405.623  ops/s
[info] ArrayOfUUIDsReading.dslJsonScala         128  thrpt    5  133557.129 ± 768.991  ops/s
[info] ArrayOfUUIDsReading.jacksonScala         128  thrpt    5   69787.906 ± 411.977  ops/s
[info] ArrayOfUUIDsReading.jsoniterScala        128  thrpt    5  245148.416 ± 859.087  ops/s
[info] ArrayOfUUIDsReading.playJson             128  thrpt    5   34075.409 ± 576.833  ops/s
[info] ArrayOfUUIDsReading.sprayJson            128  thrpt    5   16602.146 ± 102.506  ops/s
[info] ArrayOfUUIDsReading.uPickle              128  thrpt    5   42198.354 ± 113.721  ops/s
[info] ArrayOfUUIDsReading.weePickle            128  thrpt    5   46137.451 ± 198.240  ops/s
```

## After
```
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                             (size)   Mode  Cnt       Score      Error  Units
[info] ArrayOfUUIDsReading.avSystemGenCodec     128  thrpt    5   46495.502 ± 6443.539  ops/s
[info] ArrayOfUUIDsReading.borerJson            128  thrpt    5   43898.673 ±  218.682  ops/s
[info] ArrayOfUUIDsReading.circe                128  thrpt    5   43556.177 ± 2813.515  ops/s
[info] ArrayOfUUIDsReading.dslJsonScala         128  thrpt    5  134310.037 ±   80.774  ops/s
[info] ArrayOfUUIDsReading.jacksonScala         128  thrpt    5   67943.927 ± 2773.123  ops/s
[info] ArrayOfUUIDsReading.jsoniterScala        128  thrpt    5  245192.736 ± 1183.365  ops/s
[info] ArrayOfUUIDsReading.playJson             128  thrpt    5   33917.537 ±  699.567  ops/s
[info] ArrayOfUUIDsReading.sprayJson            128  thrpt    5   16902.428 ±  148.481  ops/s
[info] ArrayOfUUIDsReading.uPickle              128  thrpt    5   38186.043 ±  198.256  ops/s
[info] ArrayOfUUIDsReading.weePickle            128  thrpt    5   84668.581 ±  341.207  ops/s
```