@plokhotnyuk 2020-02-23
Thank you all for support!

Let me share my story on behind and highlight some directions for improvements.

I'm going to contribute this and [other](https://www.reddit.com/r/java/comments/darehu/jackson_release_210/f1ysf1e/) performance improvements to OpenJDK, but it is not so fast procedure these days. Also, a lot of time will flow after possible accepting of contributions until they will be used in production. So, at some time it should be revised to revert back the original `UUID.fromString` call if it will evolve to be faster in the JDK 15 release or later versions.

The next case is the version numbering: when publishing locally I've missed to add the `SNAPSHOT` suffix and overwritten my local copy of the `1.0.1` version. IMHO having this suffix be always turned on except release cutting commits will save time for all contributors here.

But the most significant case is a unit testing. This PR doesn't test parsing of upper case and non 36-byte length samples. Please consider extending number of samples, cover both positive and negative cases, taking a property based approach, and using of coverage reporting tools. Raising your testing standards and keeping testing specs in a polished gold state will help not only save on maintenance but speed up the pace of evolving your great project.