@plokhotnyuk 2020-03-06
Since build 13 improvements for `UUID.fromString` are available in OpenJDK 15 ea builds.

Bellow are results from the same laptop and the latest release of weePickle (without code from this PR). Here you can spot that improvements in this build helps to all JSON parsers:
```
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                             (size)   Mode  Cnt       Score      Error  Units
[info] ArrayOfUUIDsReading.avSystemGenCodec     128  thrpt    5   84958.904 ±  989.419  ops/s
[info] ArrayOfUUIDsReading.borerJson            128  thrpt    5   71624.892 ± 1176.907  ops/s
[info] ArrayOfUUIDsReading.circe                128  thrpt    5   82975.392 ±  283.388  ops/s
[info] ArrayOfUUIDsReading.dslJsonScala         128  thrpt    5  143260.364 ±  540.891  ops/s
[info] ArrayOfUUIDsReading.jacksonScala         128  thrpt    5   77986.303 ±  607.906  ops/s
[info] ArrayOfUUIDsReading.jsoniterScala        128  thrpt    5  294529.329 ± 1453.133  ops/s
[info] ArrayOfUUIDsReading.playJson             128  thrpt    5   52762.968 ±  324.781  ops/s
[info] ArrayOfUUIDsReading.sprayJson            128  thrpt    5   21557.913 ±   83.717  ops/s
[info] ArrayOfUUIDsReading.uPickle              128  thrpt    5   69147.195 ± 1067.182  ops/s
[info] ArrayOfUUIDsReading.weePickle            128  thrpt    5   99208.870 ±  685.534  ops/s
```